\name{SDEassignment-package}
\alias{SDEassignment-package}
\alias{SDEassignment}
\docType{package}
\title{
\packageTitle{SDEassignment}
}
\description{
\packageDescription{SDEassignment}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{SDEassignment}
\packageIndices{SDEassignment}
~~ An overview of how to use the package, including the ~~
~~ most important functions ~~
}
\author{
\packageAuthor{SDEassignment}

Maintainer: \packageMaintainer{SDEassignment}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from ~~
~~ file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
